moved isTokenPegged to psm file
